requirements.txt
# requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-filter==25.1
python-decouple==3.8
dj-database-url==2.1.0
Pillow==10.4.0
python-telegram-bot==21.3
qrcode[pil]==7.4.2
psycopg2-binary==2.9.9  # Se usar PostgreSQL